+++ back-end/test/service/subject.service.test.ts
@@ -148,7 +148,7 @@ test(`given: valid values for subject, when: subject is updated, then: subject i
     subjectDb.getSubjectById = mockSubjectDbGetSubjectById;
 
     // when
-    const result = await subjectService.updateSubject(id, newName);
+    const result = await subjectService.updateSubject(id, newName, { role: 'ADMIN' } as AuthenticationResponse);
 
     // then
     expect(result).toBeDefined();
@@ -163,7 +163,8 @@ test(`given: invalid id for subject, when: subject is updated, then: error is th
     subjectDb.getSubjectById = mockSubjectDbGetSubjectById;
 
     // when
-    const sut = async () => await subjectService.updateSubject(id, newName);
+    const sut = async () =>
+        await subjectService.updateSubject(id, newName, { role: 'ADMIN' } as AuthenticationResponse);
 
     // then
     await expect(sut).rejects.toThrowError('Subject not found');
@@ -178,11 +179,22 @@ test(`given: existing subject with new name, when: subject is updated, then: err
     subjectDb.getSubjectById = mockSubjectDbGetSubjectById;
 
     // when
-    const sut = async () => await subjectService.updateSubject(id, newName);
+    const sut = async () =>
+        await subjectService.updateSubject(id, newName, { role: 'ADMIN' } as AuthenticationResponse);
 
     // then
     await expect(sut).rejects.toThrowError('Subject already exists');
     expect(mockSubjectDbGetSubjectByName).toHaveBeenCalledWith(newName);
+});
+
+test(`given: invalid role, when: subject is updated, then: error is thrown`, async () => {
+    // given
+    // when
+    const sut = async () => await subjectService.updateSubject(id, newName, { role: 'USER' } as AuthenticationResponse);
+
+    // then
+    await expect(sut).rejects.toThrowError('Only admins can update subjects');
+    expect(mockSubjectDbUpdateSubject).not.toHaveBeenCalled();
 });
 
 test(`given: valid id for subject, when: subject is deleted, then: subject is deleted`, async () => {
@@ -193,7 +205,7 @@ test(`given: valid id for subject, when: subject is deleted, then: subject is de
     subjectDb.deleteSubject = mockSubjectDbDeleteSubject;
 
     // when
-    const result = await subjectService.deleteSubject(id);
+    const result = await subjectService.deleteSubject(id, { role: 'ADMIN' } as AuthenticationResponse);
 
     // then
     expect(result).toBeDefined();
@@ -208,9 +220,19 @@ test(`given: invalid id for subject, when: subject is deleted, then: error is th
     subjectDb.getSubjectById = mockSubjectDbGetSubjectById;
 
     // when
-    const sut = async () => await subjectService.deleteSubject(id);
+    const sut = async () => await subjectService.deleteSubject(id, { role: 'ADMIN' } as AuthenticationResponse);
 
     // then
     await expect(sut).rejects.toThrowError('Subject not found');
     expect(mockSubjectDbGetSubjectById).toHaveBeenCalledWith(id);
+});
+
+test(`given: invalid role, when: subject is deleted, then: error is thrown`, async () => {
+    // given
+    // when
+    const sut = async () => await subjectService.deleteSubject(id, { role: 'USER' } as AuthenticationResponse);
+
+    // then
+    await expect(sut).rejects.toThrowError('Only admins can delete subjects');
+    expect(mockSubjectDbDeleteSubject).not.toHaveBeenCalled();
 });
